run_make_support: add a `gcc` wrapper command
